requirements.txt
alabaster==0.7.12
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.5.2
astroid==2.11.5
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.3.0
bleach==5.0.0
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
cookiecutter==2.1.1
cryptography==37.0.2
cycler==0.11.0
DateTime==4.5
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
djangorestframework==3.14.0
docutils==0.18.1
entrypoints==0.4
executing==0.8.3
fabric==2.7.0
fake==0.8
Faker==13.13.0
fastjsonschema==2.15.3
filelock==3.8.0
flake8==4.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
greenlet==2.0.0.post0
gunicorn==20.1.0
idna==3.3
imagesize==1.3.0
inflection==0.5.1
intervaltree==3.1.0
invoke==1.7.1
ipykernel==6.13.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyring==23.6.0
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
names==0.3.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
numpydoc==1.4.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathlib2==2.3.7.post1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2022.11.5
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pylint==2.14.1
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.4.1
python-slugify==6.1.2
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.3
qstylizer==0.2.1
QtAwesome==1.1.1
qtconsole==5.3.0
QtPy==2.1.0
requests==2.28.0
rope==1.1.1
Rtree==1.0.0
Send2Trash==1.8.0
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.3.1
spyder-kernels==2.3.1
SQLAlchemy==1.4.43
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
text-unidecode==1.3
textdistance==4.2.2
three-merge==0.1.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
ujson==5.3.0
urllib3==1.26.9
uuid==1.30
virtualenv==20.16.6
virtualenv-clone==0.5.7
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.2.0
widgetsnbextension==3.6.0
wrapt==1.14.1
yapf==0.32.0
zope.interface==5.4.0
